feat: Handle extension status using events
The app now sends `checkextensionstatus` events that are catched by a
content_script from the extension. The content_script sends these events
to the extension's background. The extension background sends back its
status, and the content_script sends back an event to the app. The event
can be of two types:

* extensioninstalled: when the extension is installed, but not connected
* extensionconnected: when the extension is connected

The useExtensionStatus hook then updates its state and returns the good
value.